Pipe_Count/requirements.txt
streamlit
ultralytics>=8.1.0
gdown
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless
Pillow
numpy>=1.24.0
PyYAML
scipy
